`verify` exits immediately if `run_compile_command` fails
